Rattlingmars8/feature/edit image

+++ src/image/repository.py
@@ -1,15 +1,15 @@
-from fastapi import UploadFile, HTTPException
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.database.sql.alchemy_models import Image, User
-from src.image.schemas import ImageSchemaRequest, ImageSchemaUpdateRequest
+from src.image.schemas import ImageSchemaUpdateRequest
 
 
 class ImageQuery:
-
     @staticmethod
-    async def create(title: str, cloudinary_url: str, user: User, session: AsyncSession) -> Image:
+    async def create(
+        title: str, cloudinary_url: str, user: User, session: AsyncSession
+    ) -> Image:
         image = Image(title=title, owner_id=user.id, cloudinary_url=cloudinary_url)
         session.add(image)
         await session.commit()
@@ -22,13 +22,30 @@ class ImageQuery:
         return image.scalars().unique().one_or_none()
 
     @staticmethod
-    async def update(image: Image, image_data: ImageSchemaUpdateRequest, session: AsyncSession) -> Image:
-        image.title = image_data.title
-        await session.commit()
-        await session.refresh(image)
+    async def update(
+        image_id: int,
+        session: AsyncSession,
+        edited_cloudinary_url: str = None,
+        image_data: ImageSchemaUpdateRequest = None,
+    ) -> Image:
+        stmt = select(Image).where(Image.id == image_id)
+        image = await session.execute(stmt)
+        image = image.scalars().unique().one_or_none()
+        if image:
+            if image_data:
+                image.title = image_data.title
+            if edited_cloudinary_url:
+                image.edited_cloudinary_url = edited_cloudinary_url
+            await session.commit()
+            await session.refresh(image)
         return image
 
     @staticmethod
-    async def delete(image: Image, session: AsyncSession) -> None:
-        await session.delete(image)
-        await session.commit()
+    async def delete(image_id: int, session: AsyncSession) -> None:
+        stmt = select(Image).filter_by(id=image_id)
+        result = await session.execute(stmt)
+        image = result.scalar_one_or_none()
+        if image:
+            await session.delete(image)
+            await session.commit()
+
